refactor(discord-plays-pokemon): separate knowledge routes
